Move types to spec.go
