+++ integration/test/basic/basic_test.go
@@ -116,23 +116,6 @@ func TestChartConfigPatch(t *testing.T) {
 	}
 	log.Printf("CHARTCONFIG:: %v", chartConfig)
 
-	type Chart struct {
-		Channel string `json:"channel"`
-	}
-
-	type Metadata struct {
-		Labels map[string]string `json:"labels"`
-	}
-
-	type Spec struct {
-		Chart Chart `json:"chart"`
-	}
-
-	type ChartConfigDeployPatch struct {
-		Spec     Spec     `json:"spec"`
-		Metadata Metadata `json:"metadata"`
-	}
-
 	patch := ChartConfigDeployPatch{
 		Spec{
 			Chart: Chart{
